pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic-settings = "^2.7.1"
uvicorn = {extras = ["standart"], version = "^0.34.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
orjson = "^3.10.14"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

